requirements.txt
fastapi
uvicorn 
pillow
torch
matplotlib
python-multipart
torchvision
nltk
matplotlib
scikit-image